requirements.txt
# Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7

# LLM
openai==1.12.0

# HTTP Client
httpx==0.26.0

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0

# Dev/Test
pytest==7.4.4
pytest-asyncio==0.23.4
